# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: EduRule Semantic API
  description: Микросервис управления правилами доступа к образовательному контенту на базе OWL/SWRL.
  version: 2.1.0
servers:
  - url: http://localhost:8000/api/v1
    description: Локальный сервер разработки

paths:
  # ==========================================
  # 1 — Управление правилами (для Frontend)
  # ==========================================
  /ontology/meta:
    get:
      summary: Получить словари и метаданные онтологии
      description: Возвращает списки доступных типов правил, компетенций и статусов для отрисовки UI-редактора.
      tags: [UI Data]
      responses:
        "200":
          description: Словари успешно получены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OntologyMeta"

  /courses/{course_id}/tree:
    get:
      summary: Получить дерево курса
      description: Возвращает иерархию курса с прикрепленными политиками для компонента TreeTable.
      tags: [UI Data]
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Дерево успешно сгенерировано
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseTreeNode"

  /policies:
    get:
      summary: Получить список политик доступа
      description: Возвращает политики, отфильтрованные по ID курса или конкретного элемента.
      tags: [Policies]
      parameters:
        - name: course_id
          in: query
          required: false
          schema:
            type: string
        - name: element_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Список политик
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
    post:
      summary: Создать новое правило (AccessPolicy)
      tags: [Policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyCreate"
      responses:
        "201":
          description: Правило успешно добавлено в граф
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"

  /policies/{policy_id}:
    delete:
      summary: Удалить правило
      description: Физически удаляет узел AccessPolicy из графа (ABox).
      tags: [Policies]
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Правило успешно удалено

  /policies/{policy_id}/toggle:
    patch:
      summary: Включить/выключить правило
      description: Изменяет Data Property is_active.
      tags: [Policies]
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_active]
              properties:
                is_active:
                  type: boolean
      responses:
        "200":
          description: Статус правила изменен



  # ==========================================
  # 2 — Интеграция с СДО
  # ==========================================

  /courses/{course_id}/sync:
    post:
      summary: Синхронизация структуры курса (ABox)
      description: Загружает иерархию элементов для конкретного курса.
      tags: [Integration]
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseSyncPayload"
      responses:
        "200":
          description: Граф успешно синхронизирован
        "400":
          description: Ошибка консистентности онтологии (найдены циклы)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/progress:
    post:
      summary: Регистрация события успеваемости
      description: Webhook от СДО. Создает ProgressRecord, запускает Reasoner и обновляет кэш доступов.
      tags: [Integration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressEvent"
      responses:
        "202":
          description: Событие принято и граф пересчитан

  /access/student/{student_id}/course/{course_id}:
    get:
      summary: Получить матрицу доступов студента
      description: Возвращает массив ID образовательных элементов. Берется из In-Memory кэша.
      tags: [Integration]
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
        - name: course_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Массив доступных элементов
          content:
            application/json:
              schema:
                type: object
                properties:
                  available_elements:
                    type: array
                    items:
                      type: string

components:
  schemas:
    OntologyMeta:
      type: object
      properties:
        rule_types:
          type: array
          items:
            type: string
            example: "grade_required"
        statuses:
          type: array
          items:
            type: string
            example: "status_completed"
        competencies:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
        course_elements:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              type:
                type: string
                enum: [course, module, lecture, test]
              is_required:
                type: boolean
                default: true

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          example: "InconsistentOntologyError: Обнаружена циклическая зависимость между module_1 и module_2"

    Policy:
      type: object
      properties:
        id:
          type: string
        source_element_id:
          type: string
        rule_type:
          type: string
          enum:
            [
              grade_required,
              completion_required,
              competency_required,
              date_restricted,
            ]
        target_element_id:
          type: string
          nullable: true
        target_competency_id:
          type: string
          nullable: true
        passing_threshold:
          type: number
          nullable: true
        available_from:
          type: string
          format: date-time
          nullable: true
        available_until:
          type: string
          format: date-time
          nullable: true
        is_active:
          type: boolean
        author_id:
          type: string

    PolicyCreate:
      type: object
      required: [source_element_id, rule_type, author_id]
      properties:
        source_element_id:
          type: string
        rule_type:
          type: string
          enum:
            [
              grade_required,
              completion_required,
              competency_required,
              date_restricted,
            ]
        target_element_id:
          type: string
        target_competency_id:
          type: string
        passing_threshold:
          type: number
        available_from:
          type: string
          format: date-time
        available_until:
          type: string
          format: date-time
        author_id:
          type: string

    CourseElement:
      type: object
      required:
        - element_id
        - element_type
        - name
      properties:
        element_id:
          type: string
        element_type:
          type: string
          enum: [Module, Lecture, Test, Assignment, Practice]
        name:
          type: string
        parent_id:
          type: string
          nullable: true
        is_required:
          type: boolean
          default: true
          description: "Определяет, обязателен ли элемент для прохождения родительского модуля"
        order_index:
          type: integer
          nullable: true
          description: "Порядковый номер для сортировки. Если не передан, вычисляется неявно."

    CourseSyncPayload:
      type: object
      required: [elements]
      properties:
        elements:
          type: array
          items:
            $ref: "#/components/schemas/CourseElement"

    CourseTreeNode:
      type: object
      properties:
        key:
          type: string
        data:
          $ref: "#/components/schemas/TreeNodeData"
        children:
          type: array
          items:
            $ref: "#/components/schemas/CourseTreeNode"

    TreeNodeData:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        is_required:
          type: boolean
          description: "Обязательность элемента"
        policies:
          type: array
          items:
            $ref: "#/components/schemas/Policy"

    ProgressEvent:
      type: object
      required: [student_id, element_id, event_type]
      properties:
        student_id:
          type: string
        element_id:
          type: string
        event_type:
          type: string
          enum: [viewed, completed, graded, failed]
        grade:
          type: number
          nullable: true
        timestamp:
          type: string
          format: date-time